Refactor ONT store structure: replace ontRoot with ontRuns and update related components and tests
